fixed type conversion
string to float  was the intended conversion

+++ Embedded Code/DCDC_Micro/test.py
@@ -1,6 +1,7 @@
 import serial
 import time
 
+# change serial port
 ser = serial.Serial('COM20', 115200)
 
 
@@ -12,23 +13,23 @@ while True:
     decoded_bytes = ser_bytes[0:len(ser_bytes)-2].decode("utf-8")
     x = decoded_bytes.split(',')
 
-    Current_a = int(x[0])
-    Vin_a = int(x[1])
-    Vout_a = int(x[2])
-    Iout_a = int(x[3])
-    Temp_a = int(x[4])
-    Mode_a = int(x[5])
+    Current_a = float(x[0])
+    Vin_a = float(x[1])
+    Vout_a = float(x[2])
+    Iout_a = float(x[3])
+    Temp_a = float(x[4])
+    Mode_a = float(x[5])
 
-    Current_b = int(x[6])
-    Vin_b = int(x[7])
-    Vout_b = int(x[8])
-    Iout_b = int(x[9])
-    Temp_b = int(x[10])
-    Mode_b = int(x[11])
+    Current_b = float(x[6])
+    Vin_b = float(x[7])
+    Vout_b = float(x[8])
+    Iout_b = float(x[9])
+    Temp_b = float(x[10])
+    Mode_b = float(x[11])
 
-    Current_c = int(x[12])
-    Vin_c = int(x[13])
-    Vout_c = int(x[14])
-    Iout_c = int(x[15])
-    Temp_c = int(x[16])
-    Mode_c = int(x[17])
+    Current_c = float(x[12])
+    Vin_c = float(x[13])
+    Vout_c = float(x[14])
+    Iout_c = float(x[15])
+    Temp_c = float(x[16])
+    Mode_c = float(x[17])
